Better design for IHM, add decimal to amounts transaction

+++ application/djangoapp/models.py
@@ -31,7 +31,7 @@ class Produit(models.Model):
 
 class Transaction(models.Model):
     client_id = models.CharField(max_length=200)
-    amount = models.PositiveIntegerField()
+    amount = models.DecimalField(max_digits=10, decimal_places=2)
     date = models.DateTimeField(default=datetime.now, blank=True)
 
     def __str__(self):
